Issue #11 - React to UID Changes during ICF Update
The documentation for the ICF Update SPI very clearly states that the UID of an object can change, which is why that operation returns the updated UID. Unfortunately, though the ICF provisioner in IDM goes to great pains to ensure this updated UID gets returned to the sync operation, it appears that it gets ignored by the operation and the link object does not get updated. This issue is present in IDM 4.x and IDM 5.x (including the current release version of IDM 5.5).

This is just a band-aid patch for specifically this issue. However, the entire SyncOperation class badly needs re-factoring. The `performAction()` method alone is 175 lines and contains a nasty, double-nested switch statement with intentional fall-throughs.

Closes #11.