Introduce major refactor of openidm-repo-jdbc module
This commit adds new implementation of many openidm-repo-jdbc
components. Most components ertr written from scratch with only
a few pieces directly taken from the original code. The overall
module logic stays the same and the overall comoponent structure
as well.

To make sure the DB integration works and the refactor does not
introduce any inconsistency a testcontainer integration tests
are added. These tests can be run on the legacy components as well
as the new components. This commit intentionally leaves both
implementations so that it is possible to return to this state and
recheck legacy behavior.